requirements.txt
python-telegram-bot>=20.0
python-dotenv>=1.0
